refactor: move EnhancedDict to utils
